move system mgmt methods to dedicated files, move locking into drpc call func
